Update properties overview
