+++ docs/csharp/programming-guide/classes-and-structs/snippets/properties/TimePeriod.cs
@@ -1,0 +1,21 @@
+﻿namespace properties;
+
+// <PropertyExample>
+public class TimePeriod
+{
+    private double _seconds;
+
+    public double Hours
+    {
+        get { return _seconds / 3600; }
+        set
+        {
+            if (value < 0 || value > 24)
+                throw new ArgumentOutOfRangeException(nameof(value),
+                      "The valid range is between 0 and 24.");
+
+            _seconds = value * 3600;
+        }
+    }
+}
+// </PropertyExample>
