--- samples/snippets/csharp/programming-guide/classes-and-structs/properties-1.cs
@@ -1,33 +1,0 @@
-﻿using System;
-
-class TimePeriod
-{
-   private double _seconds;
-
-   public double Hours
-   {
-       get { return _seconds / 3600; }
-       set {
-          if (value < 0 || value > 24)
-             throw new ArgumentOutOfRangeException(
-                   $"{nameof(value)} must be between 0 and 24.");
-
-          _seconds = value * 3600;
-       }
-   }
-}
-
-class Program
-{
-   static void Main()
-   {
-       TimePeriod t = new TimePeriod();
-       // The property assignment causes the 'set' accessor to be called.
-       t.Hours = 24;
-
-       // Retrieving the property causes the 'get' accessor to be called.
-       Console.WriteLine($"Time in hours: {t.Hours}");
-   }
-}
-// The example displays the following output:
-//    Time in hours: 24
