--- samples/snippets/csharp/programming-guide/classes-and-structs/properties-2.cs
@@ -1,26 +1,0 @@
-﻿using System;
-
-public class Person
-{
-   private string _firstName;
-   private string _lastName;
-
-   public Person(string first, string last)
-   {
-      _firstName = first;
-      _lastName = last;
-   }
-
-   public string Name => $"{_firstName} {_lastName}";
-}
-
-public class Example
-{
-   public static void Main()
-   {
-      var person = new Person("Magnus", "Hedlund");
-      Console.WriteLine(person.Name);
-   }
-}
-// The example displays the following output:
-//      Magnus Hedlund
