--- samples/snippets/csharp/programming-guide/classes-and-structs/properties-3.cs
@@ -1,36 +1,0 @@
-﻿using System;
-
-public class SaleItem
-{
-   string _name;
-   decimal _cost;
-
-   public SaleItem(string name, decimal cost)
-   {
-      _name = name;
-      _cost = cost;
-   }
-
-   public string Name
-   {
-      get => _name;
-      set => _name = value;
-   }
-
-   public decimal Price
-   {
-      get => _cost;
-      set => _cost = value;
-   }
-}
-
-class Program
-{
-   static void Main(string[] args)
-   {
-      var item = new SaleItem("Shoes", 19.95m);
-      Console.WriteLine($"{item.Name}: sells for {item.Price:C2}");
-   }
-}
-// The example displays output like the following:
-//       Shoes: sells for $19.95
