--- samples/snippets/csharp/programming-guide/classes-and-structs/properties-4.cs
@@ -1,21 +1,0 @@
-﻿using System;
-
-public class SaleItem
-{
-   public string Name
-   { get; set; }
-
-   public decimal Price
-   { get; set; }
-}
-
-class Program
-{
-   static void Main(string[] args)
-   {
-      var item = new SaleItem{ Name = "Shoes", Price = 19.95m };
-      Console.WriteLine($"{item.Name}: sells for {item.Price:C2}");
-   }
-}
-// The example displays output like the following:
-//       Shoes: sells for $19.95
